Reuse ExecutorFailureTracker
