+++ core/src/main/scala/org/apache/spark/internal/config/package.scala
@@ -921,6 +921,24 @@ package object config {
       .booleanConf
       .createWithDefault(false)
 
+  private[spark] val MAX_EXECUTOR_FAILURES =
+    ConfigBuilder("spark.executor.maxNumFailures")
+      .doc("Spark exit if the number of failed executors exceeds this threshold. " +
+        "This configuration only takes affect on YARN, or on Kubernetes when " +
+        "`spark.kubernetes.allocation.pods.allocator` is set to 'direct'.")
+      .version("3.5.0")
+      .intConf
+      .createOptional
+
+  private[spark] val EXECUTOR_ATTEMPT_FAILURE_VALIDITY_INTERVAL_MS =
+    ConfigBuilder("spark.executor.failuresValidityInterval")
+      .doc("Interval after which Executor failures will be considered independent and not " +
+        "accumulate towards the attempt count. This configuration only takes affect on YARN, " +
+        "or on Kubernetes when `spark.kubernetes.allocation.pods.allocator` is set to 'direct'.")
+      .version("3.5.0")
+      .timeConf(TimeUnit.MILLISECONDS)
+      .createOptional
+
   private[spark] val UNREGISTER_OUTPUT_ON_HOST_ON_FETCH_FAILURE =
     ConfigBuilder("spark.files.fetchFailure.unRegisterOutputOnHost")
       .doc("Whether to un-register all the outputs on the host in condition that we receive " +
