+++ core/src/main/scala/org/apache/spark/util/SparkExitCode.scala
@@ -31,6 +31,9 @@ private[spark] object SparkExitCode {
   /** Exception appears when the computer cannot find the specified path. */
   val ERROR_PATH_NOT_FOUND = 3
 
+  /** Exit due to executor failures exceeds the threshold. */
+  val EXCEED_MAX_EXECUTOR_FAILURES = 11
+
   /** The default uncaught exception handler was reached. */
   val UNCAUGHT_EXCEPTION = 50
