+++ resource-managers/yarn/src/main/scala/org/apache/spark/deploy/yarn/config.scala
@@ -95,14 +95,6 @@ package object config extends Logging {
       .stringConf
       .createOptional
 
-  private[spark] val EXECUTOR_ATTEMPT_FAILURE_VALIDITY_INTERVAL_MS =
-    ConfigBuilder("spark.yarn.executor.failuresValidityInterval")
-      .doc("Interval after which Executor failures will be considered independent and not " +
-        "accumulate towards the attempt count.")
-      .version("2.0.0")
-      .timeConf(TimeUnit.MILLISECONDS)
-      .createOptional
-
   private[spark] val MAX_APP_ATTEMPTS = ConfigBuilder("spark.yarn.maxAppAttempts")
     .doc("Maximum number of AM attempts before failing the app.")
     .version("1.3.0")
@@ -277,11 +269,6 @@ package object config extends Logging {
       .version("1.2.0")
       .intConf
       .createWithDefault(25)
-
-  private[spark] val MAX_EXECUTOR_FAILURES = ConfigBuilder("spark.yarn.max.executor.failures")
-    .version("1.0.0")
-    .intConf
-    .createOptional
 
   private[spark] val MAX_REPORTER_THREAD_FAILURES =
     ConfigBuilder("spark.yarn.scheduler.reporterThread.maxFailures")
